Extract all Node.js specific child process code to `process-adapter`
