+++ lib/browser-adapter.js
@@ -1,0 +1,2 @@
+/* eslint-disable import/no-extraneous-dependencies */
+module.exports = require('ava-bundler-external/process-adapter');
