+++ lib/test-worker.js
@@ -1,39 +1,8 @@
 'use strict';
-var opts = JSON.parse(process.argv[2]);
+/* eslint-disable import-order/import-order */
+var process = require('./process-adapter');
+var opts = process.opts;
 var testPath = opts.file;
-
-// Fake TTY support
-if (opts.tty) {
-	process.stdout.isTTY = true;
-	process.stdout.columns = opts.tty.columns || 80;
-	process.stdout.rows = opts.tty.rows;
-
-	var tty = require('tty');
-	var isatty = tty.isatty;
-
-	tty.isatty = function (fd) {
-		if (fd === 1 || fd === process.stdout) {
-			return true;
-		}
-
-		return isatty(fd);
-	};
-}
-
-var path = require('path');
-var fs = require('fs');
-var debug = require('debug')('ava');
-var sourceMapSupport = require('source-map-support');
-
-if (debug.enabled) {
-	// Forward the `time-require` `--sorted` flag.
-	// Intended for internal optimization tests only.
-	if (opts._sorted) {
-		process.argv.push('--sorted');
-	}
-
-	require('time-require');
-}
 
 // bind globals first before anything has a chance to interfere
 var globals = require('./globals');
@@ -45,52 +14,20 @@ Promise.longStackTraces();
 
 (opts.require || []).forEach(require);
 
-var sourceMapCache = Object.create(null);
-
-sourceMapSupport.install({
-	environment: 'node',
-	handleUncaughtExceptions: false,
-	retrieveSourceMap: function (source) {
-		if (sourceMapCache[source]) {
-			return {
-				url: source,
-				map: fs.readFileSync(sourceMapCache[source], 'utf8')
-			};
-		}
-	}
-});
+process.installSourceMapSupport();
 
 var currentlyUnhandled = require('currently-unhandled')(); // eslint-disable-line import/order
 var serializeError = require('./serialize-error');
-var send = require('./send-to-parent');
+var send = process.send;
 var throwsHelper = require('./throws-helper');
-var installPrecompiler = require('require-precompiled'); // eslint-disable-line import/order
-var cacheDir = opts.cacheDir;
 
 // check if test files required ava and show error, when they didn't
 exports.avaRequired = false;
 
-installPrecompiler(function (filename) {
-	var precompiled = opts.precompiled[filename];
-
-	if (precompiled) {
-		sourceMapCache[filename] = path.join(cacheDir, precompiled + '.js.map');
-		return fs.readFileSync(path.join(cacheDir, precompiled + '.js'), 'utf8');
-	}
-
-	return null;
-});
+process.installPrecompilerHook();
 
 var dependencies = [];
-Object.keys(require.extensions).forEach(function (ext) {
-	var wrappedHandler = require.extensions[ext];
-	require.extensions[ext] = function (module, filename) {
-		if (filename !== testPath) {
-			dependencies.push(filename);
-		}
-		wrappedHandler(module, filename);
-	};
-});
+process.installDependencyTracking(dependencies, testPath);
 
 require(testPath);
